Fix fp warning; clear the back-buffer (for now, need it for the ui!)

+++ source/horde3d/Source/Horde3DEngine/egCom.cpp
@@ -395,7 +395,7 @@ float StatManager::getStat( int param, bool reset )
 {
 	float value;	
    float sum = 0.0;
-	int availableMem = 0;
+	float availableMem = 0;
    int integerV[4];
    int c;
